rename 'syskit-pocolog' into 'syskit-log'
Having a Pocolog namespace inside Syskit causes problems if other parts
of Syskit start to use Pocolog code (as it does for instance for
configuration logging). The problem is that once Syskit::Pocolog is
defined, it is the namespace that gets resolved instead of Pocolog
(toplevel)

Rename into 'syskit-log', which is actually closer to what it does since
it also at least handles Roby logs and text logs.