dsl: handle converting multiple streams into a single frame
This refactors the stream-to-frame conversion code to be compatible
with a stream aligner. It relies on a change in pocolog that allows
for SampleEnumerator-like functionality on DataStream itself, thus
allowing us to resample and/or select a range of a DataStream in a
way that is compatible with the stream aligner.